pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pqc-secure-transfer"
version = "1.0.0"
authors = [
    {name = "PQC Secure Transfer Contributors", email = "your-email@example.com"},
]
description = "Post-Quantum Cryptography secure data transfer system for federated learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Networking",
]
keywords = ["post-quantum", "cryptography", "federated-learning", "secure-transfer", "encryption", "quantum-safe"]
dependencies = [
    "cryptography>=41.0.0",
    "pycryptodome>=3.19.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "websockets>=12.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
full = ["liboqs-python>=0.8.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=0.18.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-benchmark>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/pqc-secure-transfer"
Documentation = "https://github.com/yourusername/pqc-secure-transfer#readme"
Repository = "https://github.com/yourusername/pqc-secure-transfer.git"
"Bug Tracker" = "https://github.com/yourusername/pqc-secure-transfer/issues"
Changelog = "https://github.com/yourusername/pqc-secure-transfer/blob/main/CHANGELOG.md"

[project.scripts]
pqc-server = "examples.server:main"
pqc-client = "examples.client:main"
pqc-demo = "simple_demo:main"

[tool.setuptools.packages.find]
include = ["pqc_secure_transfer*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["pqc_secure_transfer"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    ".tox",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "oqs.*",
    "Crypto.*",
    "tqdm.*",
    "numpy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance benchmarks",
    "security: marks tests as security-related",
]

[tool.coverage.run]
source = ["pqc_secure_transfer"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]